// frontend/src/app/core/api/generated/api/inventory.service.ts
export interface FindAllMovementsRequestParams {
    page?: number;
    size?: number;
    productId?: string;
    type?: 'INITIAL_STOCK' | 'MANUAL_IN' | 'MANUAL_OUT' | 'ORDER_RESERVED' | 'ORDER_RESERVATION_RELEASED' | 'ORDER_CONFIRMED' | 'ORDER_CANCELLED' | 'PURCHASE_RECEIVED' | 'TRANSFER_OUT' | 'TRANSFER_IN' | 'PHYSICAL_COUNT_ADJUSTMENT';
    from?: string;
    to?: string;
    reference?: string;
}

export interface FindByProductId1RequestParams {
    productId: string;
}

export interface FindLowStock1RequestParams {
    page?: number;
    size?: number;
    search?: string;
    outOfStockOnly?: boolean;
}